local-clip/pyproject.toml
[project]
name = "clip-image-classifier"
version = "0.1.0"
description = "使用CLIP模型进行零样本图像分类的Python项目"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.30.0",
    "pillow>=9.0.0",
    "matplotlib>=3.5.0",
    "requests>=2.28.0",
    "certifi>=2023.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
]

[tool.uv]
python-preference = "only-managed"